requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
selenium>=4.15.2
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1

